Make socketcluster optional
