+++ package.json
@@ -49,10 +49,10 @@
     "extendify": "^1.0.0",
     "forever": "^0.15.2",
     "glob": "^7.1.1",
-    "js-yaml": "^3.12.0",
+    "js-yaml": "^3.13.1",
     "jsonschema": "^1.2.4",
     "knex": "^0.15.2",
-    "lodash": "^4.17.10",
+    "lodash": "^4.17.11",
     "minimist": "^1.2.0",
     "moment": "^2.22.2",
     "object-mapper": "^3.0.1",
